Switch from gpgv to sopv (Stateless OpenPGP CLI Verification subset)
The Stateless OpenPGP CLI intends to be an RFC defining a common and
specified interoperable command-line interface that multiple OpenPGP
implementation provide, so that users can call any of them and switch
them transparently.

The sopv interface is a Verification only subset from SOP, that provides
the minimum necessary to perform OpenPGP verification, and has been
stabilized. In Debian there are multiple providers of this interface,
such as sqopv, rsopv, sqop, rsop, gosop and sopv-gpgv. We pick sqopv as
the default provider, and depend on the virtual package sopv.

See: https://datatracker.ietf.org/doc/draft-dkg-openpgp-stateless-cli/